Extract filter fn. for JW and BK mapping
