+++ packages/algo/quri_parts/algo/mitigation/post_selection/__init__.py
@@ -1,13 +1,9 @@
 from .post_selection import (
     PostSelectionFilterFunction,
-    create_bk_electron_number_post_selection_sampler,
     create_general_post_selection_sampler,
-    create_jw_electron_number_post_selection_sampler,
 )
 
 __all__ = [
-    "create_bk_electron_number_post_selection_sampler",
     "create_general_post_selection_sampler",
-    "create_jw_electron_number_post_selection_sampler",
     "PostSelectionFilterFunction",
 ]
